Water cluster for slice identification
The txt file contains the water coordinates belonging to some kind of water molecules cluster, identified inside the protein structure. This cluster correspond to some 3D-RISM density and could thus be useful to localize the search for density slices, to be then used to test the digital/analog algo, so to have a not too trivial (hopefully) QUBO problem.